Remove unneccessay late initializations
DO not specify the late modifier for top-level and static variables when the declaration contains an initializer.

Top-level and static variables with initializers are already evaluated lazily as if they are marked late.
So this is the opposite of what I thought but at least this mystery is solved now.
https://dart-lang.github.io/linter/lints/unnecessary_late.html